Move initialize function for third body
