+++ src/disturbances/third_body_gravity.cpp
@@ -5,6 +5,8 @@
 
 #include "third_body_gravity.hpp"
 
+#include <library/initialize/initialize_file_access.hpp>
+
 ThirdBodyGravity::ThirdBodyGravity(std::set<std::string> third_body_list, const bool is_calculation_enabled)
     : Disturbance(is_calculation_enabled, false), third_body_list_(third_body_list) {
   acceleration_i_m_s2_ = libra::Vector<3>(0.0);
@@ -52,4 +54,48 @@ std::string ThirdBodyGravity::GetLogValue() const {
   str_tmp += WriteVector(acceleration_i_m_s2_);
 
   return str_tmp;
+}
+
+ThirdBodyGravity InitThirdBodyGravity(const std::string initialize_file_path, const std::string ini_path_celes) {
+  // Generate a list of bodies to be calculated in "CelesInfo"
+  auto conf_celes = IniAccess(ini_path_celes);
+  const char* section_celes = "CELESTIAL_INFORMATION";
+  const int num_of_selected_body = conf_celes.ReadInt(section_celes, "number_of_selected_body");
+  const std::string center_object = conf_celes.ReadString(section_celes, "center_object");
+
+  std::set<std::string> selected_body_list;
+  for (int i = 0; i < num_of_selected_body; i++) {
+    std::string selected_body_id = "selected_body_name(" + std::to_string(i) + ")";
+    selected_body_list.insert(conf_celes.ReadString(section_celes, selected_body_id.c_str()));
+  }
+
+  // Generate a list of bodies to be calculated in "ThirdBodyGravity" from the list of bodies of "CelesInfo"
+  auto conf = IniAccess(initialize_file_path);
+  const char* section = "THIRD_BODY_GRAVITY";
+
+  const int num_of_third_body = conf.ReadInt(section, "number_of_third_body");
+
+  std::set<std::string> third_body_list;
+  // Generate the list of the third object if "calculation=ENABLE"
+  if (conf.ReadEnable(section, INI_CALC_LABEL)) {
+    for (int i = 0; i < num_of_third_body; i++) {
+      const std::string third_body_id = "third_body_name(" + std::to_string(i) + ")";
+      const std::string third_body_name = conf.ReadString(section, third_body_id.c_str());
+      // If the object specified by `third_body` in "SampleDisturbance.ini" is
+      // the center object of the orbital propagation, the system prints an
+      // error message.
+      assert(third_body_name != center_object);
+      // If the target specified by `third_body` in "SampleDisturbance.ini" is
+      // not in the list of bodies to be calculated by "CelesInfo", the system
+      // prints an error message.
+      assert(selected_body_list.find(third_body_name) != selected_body_list.end());
+      third_body_list.insert(third_body_name);
+    }
+  }
+
+  const bool is_calc_enable = conf.ReadEnable(section, INI_CALC_LABEL);
+  ThirdBodyGravity third_body_disturbance(third_body_list, is_calc_enable);
+  third_body_disturbance.is_log_enabled_ = conf.ReadEnable(section, INI_LOG_LABEL);
+
+  return third_body_disturbance;
 }
